Fixed several shrinks.
We now use builders whenever we can.
Builders now check for pre-conditions and return more appropriate arbitraries.
NoArbitrary is now typed to `any` instead of `never`, so it can compose seamlessly with calls that were not expected the builder to have changed the type; we are not sure that this is the best way to solve the problem, though.